Move `ViewModeSwitcher` to `Icinga\Module\Icingadb\Web\Control`
